bug fixes on serialize polymorphic belongsTo
todo:
- go back to extending base model class instead of mixins
- try moving setup/teardown into model functions
- try moving patchrunner functionality into model functions
