+++ app/components/module-sequence.js
@@ -4,9 +4,4 @@ export default ModuleGenericComponent.extend({
 
   classNames: ['module-sequence'],
 
-  init() {
-    console.log('moduleSequence init');
-    this._super(...arguments);
-  }
-
 });
