+++ app/models/module-clock.js
@@ -1,6 +1,7 @@
 import DS from 'ember-data';
 import Module from './module';
+import Patchable from '../mixins/patchable';
 
-export default Module.extend({
+export default Module.extend(Patchable, {
 
 });
